Seperate user and management SSH ports
Exactly 3 months ago I made commit 3f1a4f6eebb4a8b98ef92aec16a0d0fbfc243c06
which calls out this same issue.

Fundamentally we have a set of contradicting requirements.
A) user must have ssh with a password (for router rescue and user freedom)
B) we must have ssh that's safely possible to enable by default (eg no password auth)
C) dropbear does not allow more than one server to effectively operate on the same port
   even when on different interfaces
D) port 22 is well known but subject to many attacks.

Given this set of options the solution I have chosen is to move our management ssh
port to port 2200 with forever disabled ssh, leaving user lan ssh on a totally
independent server on port 22. Which will no longer be exposed to the outside world.

This removes the need for a toggle or any special optional behavior on user password
setting.

+++ rita/src/rita_client/dashboard/remote_access.rs
@@ -1,5 +1,4 @@
 use crate::KI;
-use crate::SETTING;
 use actix_web::http::StatusCode;
 use actix_web::HttpRequest;
 use actix_web::HttpResponse;
@@ -7,7 +6,6 @@ use actix_web::Path;
 use althea_kernel_interface::file_io::get_lines;
 use althea_kernel_interface::file_io::write_out;
 use failure::Error;
-use settings::RitaCommonSettings;
 
 static DROPBEAR_CONFIG: &str = "/etc/config/dropbear";
 static FIREWALL_CONFIG: &str = "/etc/config/firewall";
@@ -16,27 +14,42 @@ pub fn get_remote_access_status(_req: HttpRequest) -> Result<HttpResponse, Error
     if !KI.is_openwrt() {
         return Ok(HttpResponse::new(StatusCode::BAD_REQUEST));
     }
-    let lines = get_lines(DROPBEAR_CONFIG)?;
-    for line in lines.iter() {
-        if line.contains("option Interface") {
-            return Ok(HttpResponse::Ok().json(false));
-        }
-    }
-    Ok(HttpResponse::Ok().json(true))
+    Ok(HttpResponse::Ok().json(check_dropbear_config()?))
 }
 
 // todo try and combine the above function with this one and maintain
 // the http responses at some point
+#[allow(dead_code)]
 pub fn get_remote_access_internal() -> Result<bool, Error> {
     if !KI.is_openwrt() {
         return Err(format_err!("Not Openwrt!"));
     }
+    check_dropbear_config()
+}
+
+fn check_dropbear_config() -> Result<bool, Error> {
     let lines = get_lines(DROPBEAR_CONFIG)?;
+    // the old style config has one server, the new style config has two
+    // the old style has 'option interface' which indicates LAN only listening
+    // and is fine as an indicator. The new style has two blocks, the top on
+    // port 2200 and the bottom with 'option interface' to keep it constrained
+    // to lan.
     for line in lines.iter() {
-        if line.contains("option Interface") {
+        if line.contains("2200") && line.contains("option Port") {
+            // we have found our new port listening, we're good
+            return Ok(true);
+        } else if line.contains("option Interface") {
+            // since the new config has port 2200 above option interface
+            // in the second server directive this means we are not enabled
+            // in the new config. Likewise the old config doesn't have the
+            // interface directive at all. Either way if we hit this remote
+            // access is not enabled
             return Ok(false);
         }
     }
+    // we have found neither the new configs listening directive, nor the old configs
+    // not listening line, this means we're in the old config and have remote access
+    // enabled
     Ok(true)
 }
 
@@ -48,34 +61,56 @@ pub fn set_remote_access_status(path: Path<bool>) -> Result<HttpResponse, Error>
 
 pub fn set_remote_access_internal(remote_access: bool) -> Result<(), Error> {
     let mut lines: Vec<String> = Vec::new();
+    // the wonky spacing is actually important, keep it around.
+    // dropbear server one is ours for remote access, it never allows password
+    // auth and listens on the higher port 2200
+    if remote_access {
+        lines.push("config dropbear".to_string());
+        lines.push("        option PasswordAuth 'no'".to_string());
+        lines.push("        option Port         '2200'".to_string());
+    }
+    // password auth is enabled by default
     lines.push("config dropbear".to_string());
-    // the wonky spacing is actually important, keep it
-    if !remote_access || SETTING.get_network().rita_dashboard_password.is_some() {
-        lines.push("        option PasswordAuth 'yes'".to_string());
-    }
     lines.push("        option Port         '22'".to_string());
-    if !remote_access {
-        lines.push("        option Interface    'lan'".to_string());
-    }
+    lines.push("        option Interface    'lan'".to_string());
 
     write_out(DROPBEAR_CONFIG, lines)?;
     KI.run_command("/etc/init.d/dropbear", &["restart"])?;
 
-    // this should just be cruft now as we don't ship any version
-    // without this in the firewall config, but leaving it here just in case.
+    // this adds the updated rules to the firewall config, notice the versioning on the
+    // firewall rules. The old ones will be left in place.
     let mut firewall_lines = get_lines(FIREWALL_CONFIG)?;
+    let mut needs_mesh_ssh = true;
+    let mut needs_wan_ssh = true;
     for line in firewall_lines.iter() {
-        if line.contains("Allow-Mesh-SSH") {
-            return Ok(());
+        if line.contains("Allow-Mesh-SSH-2") {
+            needs_mesh_ssh = false;
+        }
+        if line.contains("Allow-WAN-SSH") {
+            needs_wan_ssh = false;
         }
     }
-    firewall_lines.push("".to_string());
-    firewall_lines.push("config rule".to_string());
-    firewall_lines.push("        option name             Allow-Mesh-SSH".to_string());
-    firewall_lines.push("        option src              mesh".to_string());
-    firewall_lines.push("        option dest_port        22".to_string());
-    firewall_lines.push("        option target           ACCEPT".to_string());
-    write_out(FIREWALL_CONFIG, firewall_lines)?;
-    KI.run_command("reboot", &[])?;
-    Ok(())
+    if needs_mesh_ssh {
+        firewall_lines.push("".to_string());
+        firewall_lines.push("config rule".to_string());
+        firewall_lines.push("        option name             Allow-Mesh-SSH-2".to_string());
+        firewall_lines.push("        option src              mesh".to_string());
+        firewall_lines.push("        option dest_port        2200".to_string());
+        firewall_lines.push("        option target           ACCEPT".to_string());
+    }
+    if needs_wan_ssh {
+        firewall_lines.push("".to_string());
+        firewall_lines.push("config rule".to_string());
+        firewall_lines.push("        option name             Allow-WAN-SSH".to_string());
+        firewall_lines.push("        option src              backhaul".to_string());
+        firewall_lines.push("        option dest_port        2200".to_string());
+        firewall_lines.push("        option target           ACCEPT".to_string());
+    }
+    if needs_mesh_ssh || needs_wan_ssh {
+        write_out(FIREWALL_CONFIG, firewall_lines)?;
+        KI.run_command("reboot", &[])?;
+        Ok(())
+    } else {
+        Ok(())
+    }
 }
